+++ checkpoint/orbax/checkpoint/experimental/v1/_src/handlers/composite_handler.py
@@ -27,6 +27,7 @@ from orbax.checkpoint._src.metadata import step_metadata_serialization
 from orbax.checkpoint.experimental.v1._src.context import context as context_lib
 from orbax.checkpoint.experimental.v1._src.handlers import registration
 import orbax.checkpoint.experimental.v1._src.handlers.global_registration  # pylint: disable=unused-import
+from orbax.checkpoint.experimental.v1._src.path import format_utils
 from orbax.checkpoint.experimental.v1._src.path import types as path_types
 
 
@@ -194,8 +195,7 @@ class CompositeHandler:
     return loadable_checkpointable_names_to_handlers
 
   def _get_saved_handler_typestrs(
-      self,
-      directory: path_types.Path,
+      self, directory: path_types.Path
   ) -> dict[str, str]:
     """Reads from the checkpoint metadata to get saved handler typestrs."""
     step_metadata_file_path = checkpoint_metadata.step_metadata_file_path(
@@ -206,19 +206,27 @@ class CompositeHandler:
       saved_metadata = step_metadata_serialization.deserialize(
           serialized_metadata or {}
       )
-      assert isinstance(saved_metadata.item_handlers, dict)
-      return saved_metadata.item_handlers
+      if isinstance(saved_metadata.item_handlers, dict):
+        return saved_metadata.item_handlers  # found step level metadata.
+      raise ValueError(
+          'Expected a valid path containing checkpointable subdirectories, but'
+          ' given path contains subdirectories:'
+          f' {format_utils.subdirs(directory)}... Given path is {directory}.'
+          ' _CHECKPOINT_METADATA file under given path has'
+          f' `item_handlers`={saved_metadata.item_handlers}, whose keys should'
+          ' match the checkpointable subdirectory names. If you intended to'
+          ' load a pytree checkpoint from the given path, then please consider'
+          ' using `loading.load_pytree(..., checkpointable_name=None)` instead.'
+      )
 
     logging.warning(
-        'Given dir contains checkpointables subdirs but no step metadata'
-        ' file=%s. Such dirs can exist if the checkpoints are saved directly'
-        ' using V0 Checkpointer instead of using CheckpointManager or'
-        ' CompositeCheckpointHandler. Will fetch saved handlers from each of'
-        ' the checkpointable subdirectories.',
+        'Given dir does not contain checkpoint metadata file: %s. Trying to get'
+        ' saved handlers from checkpoint metadata in each of the checkpointable'
+        ' subdirectory.',
         directory,
     )
 
-    saved_handler_typestrs = {}
+    saved_handler_typestrs: dict[str, str] = {}
     for checkpointable_path in directory.iterdir():
       serialized_metadata = self._metadata_store.read(
           checkpoint_metadata.step_metadata_file_path(checkpointable_path)
@@ -228,7 +236,17 @@ class CompositeHandler:
       saved_metadata = step_metadata_serialization.deserialize(
           serialized_metadata
       )
-      assert not isinstance(saved_metadata.item_handlers, dict)
+      if isinstance(saved_metadata.item_handlers, dict):
+        raise ValueError(
+            'Expected a valid path containing checkpointable subdirectories,'
+            ' but given path contains subdirectories:'
+            f' {format_utils.subdirs(directory)}... Given path is {directory}.'
+            ' _CHECKPOINT_METADATA file under a subdir of given path has'
+            f' `item_handlers`={saved_metadata.item_handlers}, whose keys'
+            ' should match the checkpointable subdirectory names. Did you mean'
+            ' to provide the following subdirectory path instead:'
+            f' {checkpointable_path}?'
+        )
       item_handlers = saved_metadata.item_handlers
       if item_handlers is not None:
         checkpointable_name = checkpointable_path.name